+++ posthog/src/main/java/com/posthog/PostHog.kt
@@ -1325,6 +1325,16 @@ public class PostHog private constructor(
         return sessionReplayHandler?.isActive() == true && isSessionActive()
     }
 
+    private fun shouldRecordSession(): Boolean {
+        val sessionId = PostHogSessionManager.getActiveSessionId()?.toString()
+        if (sessionId != null) {
+            // sampling is deterministic so we can sample again for the same session id
+            return remoteConfig?.makeSamplingDecision(sessionId) ?: true
+        }
+        // no session id
+        return false
+    }
+
     override fun startSessionReplay(resumeCurrent: Boolean) {
         if (!isEnabled()) {
             return
@@ -1344,10 +1354,19 @@ public class PostHog private constructor(
             }
 
             if (resumeCurrent) {
+                if (!shouldRecordSession()) {
+                    return
+                }
+
                 it.start(true)
             } else {
                 endSession()
                 startSession()
+
+                if (!shouldRecordSession()) {
+                    return
+                }
+
                 it.start(false)
             }
         } ?: run {